tests/wpt/web-platform-tests/media-source/URL-createObjectURL-null.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>URL.createObjectURL(null)</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
test(function() {
    assert_throws_js(TypeError, function() {
        window.URL.createObjectURL(null);
    });
}, "URL.createObjectURL(null)");
</script>
</body>
</html>
